refactor: reorganize conjugation tests and modularize focus-specific logic
