+++ shared/utils/word/conjugateMag.ts
@@ -1,0 +1,22 @@
+import type { Aspect } from './conjugate.ts'
+import { attachPrefix, reduplicate } from './conjugateHelpers.ts'
+
+/**
+ * MAG focus:
+ *   inf: mag + root
+ *   comp: nag + root
+ *   incomp: nag + redup(root)
+ *   cont: mag + redup(root)
+ */
+export function conjMAG(root: string, aspect: Aspect): string {
+	switch (aspect) {
+		case 'infinitive':
+			return attachPrefix('mag', root)
+		case 'completed':
+			return attachPrefix('nag', root)
+		case 'incompleted':
+			return attachPrefix('nag', reduplicate(root))
+		case 'contemplated':
+			return attachPrefix('mag', reduplicate(root))
+	}
+}
