+++ shared/utils/word/conjugateUm.ts
@@ -1,0 +1,22 @@
+import type { Aspect } from './conjugate.ts'
+import { insertInfix, reduplicate } from './conjugateHelpers.ts'
+
+/**
+ * UM focus (actor):
+ *   inf/comp: insert -um- after first consonant (or prefix if vowel-initial)
+ *   incomp:   insert -um- and reduplicate the syllable after it (practical shortcut: reinsert "um" then add first syllable)
+ *   cont:     redup(root) (no "um" in future)
+ *
+ * Note: For many roots like "luto", UM exists (lumuto/lumuluto/luluto) but MAG is far more idiomatic.
+ */
+export function conjUM(root: string, aspect: Aspect): string {
+	switch (aspect) {
+		case 'infinitive':
+		case 'completed':
+			return insertInfix(root, 'um')
+		case 'incompleted':
+			return insertInfix(reduplicate(root), 'um')
+		case 'contemplated':
+			return reduplicate(root) // kakain, luluto, iinom
+	}
+}
